Create different refiner instances from a single class
