--- src/refiners/TimeFormatRefiner.ts
@@ -1,34 +1,0 @@
-import {
-	ParsedResult,
-	IRefiner,
-} from '../types';
-
-class TimeFormatRefiner implements IRefiner {
-	public readonly name: string;
-
-	constructor(name: string) {
-		this.name = name;
-	}
-
-	refine(parsedResults: Array<ParsedResult>): Array<ParsedResult> {
-		parsedResults.forEach(r => {
-			let meridiemExists = false;
-			r.tokens.forEach(t => {
-				if (t.type === 'meridiem') {
-					meridiemExists = true;
-				}
-			});
-			if (meridiemExists) {
-				r.tokens.forEach(t => {
-					if (t.type === 'twentyFourHour') {
-						t.setType('twelveHour');
-					}
-				});
-			}
-		});
-
-		return parsedResults;
-	}
-}
-
-export default TimeFormatRefiner;
